gitlab-cd: add deb_delivery role (copy and install packages)
Only active if delivery_deb_files is defined and not empty. Existing
playbook may bot be affected by the change if igloo-ansible-roles is
up-to-date.

playbook_sudoers_deb must be true to provision sudoers rules (utils role)

If delivery_deb_files variables is not empty and
playbook_sudoers_deb not true, an error is thrown on
delivery.

This change is safe for existing projets (no delivery_deb_files, playbook_sudoers_deb
provided).